Fix critical build warnings and bugs
Critical Fixes:
- Fix wrong function call in geometric_recovery_orchestrator.c (convergence_detector_create_custom)
- Fix type incompatibility warnings in hierarchical_threading.c
- Fix format string mismatches in anchor_grid.c (%u -> %lu for uint64_t)

Code Quality Improvements:
- Remove _GNU_SOURCE redefinitions (4 files)
- Clean up unused variables (6 fixes)
- Mark unused parameters appropriately (5 fixes)
- Add TODO comments for future implementation

Build Statistics:
- Reduced warnings from 70+ to 57 (18% reduction)
- Fixed 1 critical bug
- Fixed all high-priority warnings
- Build completes successfully

Verified:
- All libraries build successfully
- No compilation errors
- src/ directory not needed for build (only Python bindings)

+++ algorithms/src/geometric_recovery/anchor_grid.c
@@ -533,8 +533,8 @@ void anchor_grid_print_stats(const AnchorGrid* grid) {
     
     if (grid->solid) {
         printf("\nPlatonic Solid:\n");
-        printf("  Edges: %u\n", grid->solid->num_edges);
-        printf("  Faces: %u\n", grid->solid->num_faces);
+        printf("  Edges: %lu\n", (unsigned long)grid->solid->num_edges);
+        printf("  Faces: %lu\n", (unsigned long)grid->solid->num_faces);
     }
     
     printf("\n12-Fold Symmetry: %s\n",
